fix recursive_dijkstra
Storable is unable to freeze a structure that contains a coderef.
What doesn't make sense is that it can't deepclone one either.
So the coderef that is in our to-be-cloned structure is saved to
a temporary variable, removed from the structure, the clone done and
the ref written back...

Since recursive_dijkstra is a hack anyway this change shall feel
welcome there. :)